Add CC BY-licensed images (see commit msg)
This application includes the following images that, as of the
date and time of this commit message, were released under the
Creative Commons Attribution (CC BY) license.

These images are included in this open source software with
attribution, and a link to the original work, in compliance
with Creative Commons Attribution license under which they
were released/obtained. The following modifications were made
to each of the images:
  - resized each image to 1920px by 1080px (240 ppi)

List of licensed works:
  - "Day 105: Oregon Coast Range" by Loren Kerns on Flickr
    - https://www.flickr.com/photos/lorenkerns/8651732785

  - "Mt. Hood, Oregon" by tsaiproject on Flickr
    - https://www.flickr.com/photos/tsaiproject/9943809254

  - "Multnomah Falls" by John Tregoning on Flickr
    - https://www.flickr.com/photos/johnniewalker/12660211844

  - "101_1812" by Randall Booth on Flickr
    - https://www.flickr.com/photos/randall-booth/9060319329

  - "Central Oregon Landscape" by Michael McCullough on Flickr
    - https://www.flickr.com/photos/ex_magician/3196286183

  - "Oregon Autumn Part 4" by Ian Sane on Flickr
    - https://www.flickr.com/photos/31246066@N04/4030400633

  - "Astoria, Oregon" by Linda Tanner on Flickr
    - https://www.flickr.com/photos/goingslo/11522920406