use glyphicons instead of unicode chars
Glyphicons are much cleaner and easier to see from a distance than the
unicode characters I originally used. At the time of this commit,
Glyphicons were free to use with attribution, under the Creative
Commons Attribution 3.0 Unported (CC BY 3.0) license.

The following files are from glyphicons.com and were modified to be
white in color instead of black:
    - public/img/cloud.png
    - public/img/sun.png
    - public/img/umbrella.png