implement views, closes gh-5
This series of changes implements "views" in Waylon. Each view can have
one or more Jenkins servers attached to it, and each server can be
configured to query one or more jobs.

This introduces navigation in the page header for each view, and changes
the landing page to display a list for the user to select from.