use jquery for data refreshes and nirvana mode
Prior to this commit, data was refreshed by a meta refresh reloading the
entire page. For obvious reasons, that's inefficient, and ugly (the user
can tell the page is being refreshed). We're now making use of jQuery to
handle interactions and display changes elegantly.

The changes in this commit achieve the following:
  - the data (via HTML table) in its own route (`/view/:name/data`)
  - on the first load of the page, message and animated graphic
  - jQuery refreshes every `refresh_interval` seconds
  - the data just "appears"; the user doesn't see a page refresh
  - use jQuery for entering and exiting nirvana mode