'Under Investigation' buttons, display-only. gh-3
This is the beginning of the implementation of GitHub issue #3,
implementing an 'under investigation' feature. Right now, this only
queries the last build for a failed job to see if the description has
been set to `/under investigation/i`. If it has, display a disabled
button 'Under investigation'. If it hasn't, display a clickable button
'Investigate'.