mark failed jobs as 'under investigation' gh-3
This is the second commit for GitHub #3, "Ability to mark a failed
job/build as 'under investigation'". At a high level, it implements the
following:
- Authenticates to a Jenkins instance
- POSTs to the API to set the description on the failed build
- Open the failed build in a new tab

I'm not marking this feature complete just yet; it doesn't yet have a
lot of the error checking that other routes do, and I'd really like to
clean up some of the code into helper functions. But... it works :)