better handling of non-existent jobs. fixes gh-6.
If Waylon queries a Jenkins server for a non-existent job, (the job
could have been deleted, a typo in the config file, anything really),
`JenkinsApi::Exceptions::NotFound` is raised. This commit rescues this
exception and displays a warning to the user.

Furthermore, we enhance nirvana mode by ensuring at least one known good
element (e.g. `.successful-job`) is present on the page.