build progress indicator and est. time remaining
This commit implements gh-9, adding a progress bar to jobs that are
building, and provides an estimated time remaining. These values are
each calculated by Jenkins and just pulled via the API.

If a job is "stuck" or otherwise the API call returns garbage, the
progress bar will turn red, and the est. time remaining will change to
infinity.